chore(tui): split runtime task completion
